feat(+page.svelte): add "show more" button to display additional favorite colors
The "+page.svelte" file has been modified to include a new button that allows users to show more favorite colors. The button is displayed when the number of favorite colors exceeds the maximum limit set by the "favoritesMax" variable. When clicked, the button increases the "favoritesMax" value by 10, allowing more favorite colors to be shown.